remove docoding mod
